+++ scanpipe/views.py
@@ -33,6 +33,7 @@ from django.conf import settings
 from django.contrib import messages
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.core.exceptions import SuspiciousFileOperation
+from django.core.exceptions import ValidationError
 from django.core.files.storage.filesystem import FileSystemStorage
 from django.db.models import Prefetch
 from django.db.models.manager import Manager
@@ -47,6 +48,7 @@ from django.shortcuts import render
 from django.template.defaultfilters import filesizeformat
 from django.urls import reverse
 from django.urls import reverse_lazy
+from django.utils.decorators import method_decorator
 from django.views import generic
 from django.views.decorators.http import require_POST
 from django.views.generic.detail import SingleObjectMixin
@@ -534,6 +536,11 @@ class ProjectListView(
         },
     ]
 
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["archive_form"] = ArchiveProjectForm()
+        return context
+
     def get_queryset(self):
         return (
             super()
@@ -669,7 +676,6 @@ class ProjectDetailView(ConditionalLoginRequired, generic.DetailView):
                 "add_inputs_form": AddInputsForm(),
                 "add_labels_form": AddLabelsForm(),
                 "project_clone_form": ProjectCloneForm(project),
-                "archive_form": ArchiveProjectForm(),
                 "project_resources_url": project_resources_url,
                 "resource_status_summary": resource_status_summary,
                 "resource_licenses_summary": resource_licenses_summary,
@@ -732,6 +738,11 @@ class ProjectSettingsView(ConditionalLoginRequired, UpdateView):
         if request.GET.get("download"):
             return self.download_config_file(project=self.get_object())
         return super().get(request, *args, **kwargs)
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["archive_form"] = ArchiveProjectForm()
+        return context
 
     @staticmethod
     def download_config_file(project):
@@ -901,11 +912,7 @@ class ProjectArchiveView(ConditionalLoginRequired, SingleObjectMixin, FormView):
 
         project = self.get_object()
         try:
-            project.archive(
-                remove_input=form.cleaned_data["remove_input"],
-                remove_codebase=form.cleaned_data["remove_codebase"],
-                remove_output=form.cleaned_data["remove_output"],
-            )
+            project.archive(**form.cleaned_data)
         except RunInProgressError as error:
             messages.error(self.request, error)
             return redirect(project)
@@ -929,6 +936,57 @@ class ProjectDeleteView(ConditionalLoginRequired, generic.DeleteView):
 
         messages.success(self.request, self.success_message.format(project.name))
         return response_redirect
+
+
+@method_decorator(require_POST, name="dispatch")
+class ProjectActionView(ConditionalLoginRequired, generic.ListView):
+    """Call a method for each instance of the selection."""
+
+    model = Project
+    allowed_actions = ["archive", "delete", "reset"]
+    success_url = reverse_lazy("project_list")
+
+    def post(self, request, *args, **kwargs):
+        action = request.POST.get("action")
+        if action not in self.allowed_actions:
+            raise Http404
+
+        selected_ids = request.POST.get("selected_ids", "").split(",")
+        count = 0
+
+        action_kwargs = {}
+        if action == "archive":
+            archive_form = ArchiveProjectForm(request.POST)
+            if not archive_form.is_valid():
+                raise Http404
+            action_kwargs = archive_form.cleaned_data
+
+        for project_uuid in selected_ids:
+            if self.perform_action(action, project_uuid, action_kwargs):
+                count += 1
+
+        if count:
+            messages.success(self.request, self.get_success_message(action, count))
+
+        return HttpResponseRedirect(self.success_url)
+
+    def perform_action(self, action, project_uuid, action_kwargs=None):
+        if not action_kwargs:
+            action_kwargs = {}
+
+        try:
+            project = Project.objects.get(pk=project_uuid)
+            getattr(project, action)(**action_kwargs)
+            return True
+        except Project.DoesNotExist:
+            messages.error(self.request, f"Project {project_uuid} does not exist.")
+        except RunInProgressError as error:
+            messages.error(self.request, str(error))
+        except (AttributeError, ValidationError):
+            raise Http404
+
+    def get_success_message(self, action, count):
+        return f"{count} projects have been {action}."
 
 
 class ProjectResetView(ConditionalLoginRequired, generic.DeleteView):